+++ ClassData/GlobalConfig.cs
@@ -24,7 +24,7 @@ namespace NASARData
     /// </summary>
     public class GlobalConfig
     {
-        public static readonly string ProgramVersion = "0.7.6";
+        public static readonly string ProgramVersion = "0.7.7";
 
         public static string GithubVersion = "";
 
@@ -67,6 +67,16 @@ namespace NASARData
 
         // Temp path for the user. ie: C:\Users\nik\AppData\Local\Temp\NASR_TO_SCT
         public static readonly string tempPath = $"{Path.GetTempPath()}NASR2SCT";
+
+        public static List<MetaAirportModel> AllMetaFileAirports = new List<MetaAirportModel>();
+
+        public static string facilityID;
+
+        public static List<string> allArtcc = new List<string>() {
+            "FAA","FIM","SBA","ZAB","ZAK","ZAN","ZAP","ZAU","ZBW","ZDC","ZDV","ZEG","ZFW","ZHN",
+            "ZHU","ZID","ZJX","ZKC","ZLA","ZLC","ZMA","ZME","ZMP","ZNY","ZOA","ZOB","ZSE","ZSU",
+            "ZTL","ZUA","ZUL","ZVR","ZWG","ZYZ"
+        };
 
         public static bool GetMetaUrlResponse() 
         {
@@ -130,6 +140,7 @@ namespace NASARData
 
         public static void DownloadAllFiles(string effectiveDate, string airacCycle)
         {
+            DownloadedFilePaths = new List<string>();
             Dictionary<string, string> allURLs = new Dictionary<string, string>()
             {
                 { $"{effectiveDate}_STARDP.zip", $"https://nfdc.faa.gov/webContent/28DaySub/{effectiveDate}/STARDP.zip" },
@@ -140,7 +151,8 @@ namespace NASARData
                 { $"{effectiveDate}_AWY.zip", $"https://nfdc.faa.gov/webContent/28DaySub/{effectiveDate}/AWY.zip"},
                 { $"{airacCycle}_FAA_Meta.xml", $"https://aeronav.faa.gov/d-tpp/{airacCycle}/xml_data/d-tpp_Metafile.xml"},
                 { $"{effectiveDate}_FIX.zip", $"https://nfdc.faa.gov/webContent/28DaySub/{effectiveDate}/FIX.zip" },
-                { $"{effectiveDate}_NAV.zip", $"https://nfdc.faa.gov/webContent/28DaySub/{effectiveDate}/NAV.zip"}
+                { $"{effectiveDate}_NAV.zip", $"https://nfdc.faa.gov/webContent/28DaySub/{effectiveDate}/NAV.zip"},
+                { $"{airacCycle}_TELEPHONY.html", $"https://www.faa.gov/air_traffic/publications/atpubs/cnt_html/chap3_section_2.html" }
             };
 
             // Web Client used to connect to the FAA website.